+++ src/schemes/fluid/entropically_damped_sph/system.jl
@@ -122,9 +122,10 @@ function EntropicallyDampedSPHSystem(initial_condition; smoothing_kernel, smooth
     if surface_tension !== nothing && surface_normal_method === nothing
         surface_normal_method = ColorfieldSurfaceNormal()
     end
+    validate_corrected_csf(surface_normal_method, surface_tension)
 
     if surface_normal_method !== nothing && reference_particle_spacing < eps()
-        throw(ArgumentError("`reference_particle_spacing` must be set to a positive value when using `ColorfieldSurfaceNormal` or a surface tension model"))
+        throw(ArgumentError("`reference_particle_spacing` must be set to a positive value when using a surface-normal method or a surface tension model"))
     end
 
     if correction isa ShepardKernelCorrection &&