+++ src/schemes/fluid/weakly_compressible_sph/system.jl
@@ -133,9 +133,10 @@ function WeaklyCompressibleSPHSystem(initial_condition; smoothing_kernel,
     if surface_tension !== nothing && surface_normal_method === nothing
         surface_normal_method = ColorfieldSurfaceNormal()
     end
+    validate_corrected_csf(surface_normal_method, surface_tension)
 
     if surface_normal_method !== nothing && reference_particle_spacing < eps()
-        throw(ArgumentError("`reference_particle_spacing` must be set to a positive value when using `ColorfieldSurfaceNormal` or a surface tension model"))
+        throw(ArgumentError("`reference_particle_spacing` must be set to a positive value when using a surface-normal method or a surface tension model"))
     end
 
     pressure_acceleration = choose_pressure_acceleration_formulation(pressure_acceleration,